+++ utils/build_swift/tests/argparse/test_parser.py
@@ -1,6 +1,6 @@
 # This source file is part of the Swift.org open source project
 #
-# Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
+# Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 # Licensed under Apache License v2.0 with Runtime Library Exception
 #
 # See https://swift.org/LICENSE.txt for license information
@@ -10,7 +10,7 @@
 from argparse import _ArgumentGroup, _MutuallyExclusiveGroup
 
 from ..utils import TestCase
-from ...argparse import ArgumentParser, actions
+from ...build_swift.argparse import ArgumentParser, actions
 
 
 # -----------------------------------------------------------------------------
