+++ utils/build_swift/tests/expected_options.py
@@ -1,6 +1,6 @@
 # This source file is part of the Swift.org open source project
 #
-# Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
+# Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 # Licensed under Apache License v2.0 with Runtime Library Exception
 #
 # See https://swift.org/LICENSE.txt for license information
@@ -12,8 +12,8 @@ import multiprocessing
 from swift_build_support.swift_build_support import host
 from swift_build_support.swift_build_support import targets
 
-from .. import argparse
-from .. import defaults
+from ..build_swift import argparse
+from ..build_swift import defaults
 
 
 __all__ = [
