+++ utils/build_swift/tests/test_migration.py
@@ -1,6 +1,6 @@
 # This source file is part of the Swift.org open source project
 #
-# Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
+# Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 # Licensed under Apache License v2.0 with Runtime Library Exception
 #
 # See https://swift.org/LICENSE.txt for license information
@@ -10,7 +10,7 @@
 from __future__ import absolute_import, unicode_literals
 
 from .utils import TestCase, add_metaclass
-from .. import migration
+from ..build_swift import migration
 
 
 def _get_sdk_targets(sdk_names):
